+++ python/libcasm/enumerate/__init__.py
@@ -5,6 +5,11 @@ from ._ConfigEnumAllOccupations import (
 from ._ConfigEnumInfo import (
     ConfigEnumInfo,
 )
+from ._ConfigEnumLocalOccupations import (
+    ConfigEnumLocalOccupations,
+    LocalConfiguration,
+    LocalConfigurationList,
+)
 from ._ConfigEnumMeshGrid import (
     ConfigEnumMeshGrid,
     irreducible_wedge_points,
@@ -22,7 +27,17 @@ from ._enumerate import (
 from ._methods import (
     make_all_distinct_local_perturbations,
     make_all_distinct_periodic_perturbations,
+    make_equivalents_generators,
+    make_first_n_orbits,
+    make_first_n_orbits_cluster_specs,
+    make_occevent_equivalents_generators,
     make_occevent_suborbits,
+)
+from ._OccEventPrimSymInfo import (
+    OccEventPrimSymInfo,
+)
+from ._OccEventSupercellSymInfo import (
+    OccEventSupercellSymInfo,
 )
 from ._ScelEnum import (
     ScelEnum,